+++ requirements.txt
@@ -1,0 +1,3 @@
+setuptools~=58.1.0
+numpy~=1.23.4
+matplotlib~=3.6.1
